old-tests/submission/W3C/domhtml/HTMLHeadElement01.html

<!DOCTYPE html>
<html>
<head profile="http://www.w3.org/2004/07/profile">
<meta charset="utf-8">
<title>NIST DOM HTML Test - HEAD</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-head-profile" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>Hello, World.</p>



<script>
var domtest = async_test("profile IDL attribute shoudn't be present");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vprofile;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("head");
      assert_equals(nodeList.length,1);
      testNode = nodeList.item(0);
      vprofile = testNode.profile;
      assert_not_equals(vprofile, "http://www.w3.org/2004/07/profile");

    domtest.done();
}));

</script>
<p>The profile IDL attribute on head elements (with the HTMLHeadElement interface) is intentionally omitted. Unless so required by another applicable specification, implementations would therefore not support this attribute. (It is mentioned here as it was defined in a previous version of the DOM specifications.)</p>
<div id="log"></div>
</body></html>
